# frontend/openapi.json
{
  "openapi": "3.1.0",
  "info": {
    "title": "AWA Platform",
    "version": "0.1.0"
  },
  "paths": {
...
    "/inspectors": {
      "get": {
        "tags": [
          "inspectors"
        ],
        "summary": "List Inspectors",
        "operationId": "list_inspectors_inspectors_get",
        "parameters": [
          {
            "name": "state",
            "in": "query",
            "required": false,
            "schema": {
              "anyOf": [
                {
                  "type": "string"
                },
                {
                  "type": "null"
                }
              ],
              "title": "State"
            }
          },
          {
            "name": "limit",
            "in": "query",
            "required": false,
            "schema": {
              "type": "integer",
              "maximum": 100,
              "minimum": 1,
              "default": 20,
              "title": "Limit"
            }
          },
          {
            "name": "offset",
            "in": "query",
            "required": false,
            "schema": {
              "type": "integer",
              "minimum": 0,
              "default": 0,
              "title": "Offset"
            }
          },
          {
            "name": "x-api-key",
            "in": "header",
            "required": false,
            "schema": {
              "anyOf": [
                {
                  "type": "string"
                },
                {
                  "type": "null"
                }
              ],
              "title": "X-Api-Key"
            }
          }
        ],
        "responses": {
          "200": {
            "description": "Successful Response",
            "content": {
              "application/json": {
                "schema": {
                  "$ref": "#/components/schemas/InspectorListOut"
                }
              }
            }
          },
          "422": {
            "description": "Validation Error",
            "content": {
              "application/json": {
                "schema": {
                  "$ref": "#/components/schemas/HTTPValidationError"
                }
              }
            }
          }
        }
      }
    },
    "/inspectors/{inspector_id}": {
      "get": {
        "tags": [
          "inspectors"
        ],
        "summary": "Get Inspector",
        "operationId": "get_inspector_inspectors__inspector_id__get",
        "parameters": [
          {
            "name": "inspector_id",
            "in": "path",
            "required": true,
            "schema": {
              "type": "string",
              "title": "Inspector Id"
            }
          },
          {
            "name": "x-api-key",
            "in": "header",
            "required": false,
            "schema": {
              "anyOf": [
                {
                  "type": "string"
                },
                {
                  "type": "null"
                }
              ],
              "title": "X-Api-Key"
            }
          }
        ],
        "responses": {
          "200": {
            "description": "Successful Response",
            "content": {
              "application/json": {
                "schema": {
                  "$ref": "#/components/schemas/InspectorDetailOut"
                }
              }
            }
          },
          "422": {
            "description": "Validation Error",
            "content": {
              "application/json": {
                "schema": {
                  "$ref": "#/components/schemas/HTTPValidationError"
                }
              }
            }
          }
        }
      }
    },
...
  },
  "components": {
    "schemas": {
...
      "HTTPValidationError": {
        "properties": {
          "detail": {
            "items": {
              "$ref": "#/components/schemas/ValidationError"
            },
            "type": "array",
            "title": "Detail"
          }
        },
        "type": "object",
        "title": "HTTPValidationError"
      },
      "InspectionOut": {
        "properties": {
          "id": {
            "type": "integer",
            "title": "Id"
          },
          "facility_id": {
            "anyOf": [
              {
                "type": "integer"
              },
              {
                "type": "null"
              }
            ],
            "title": "Facility Id"
          },
          "facility_name": {
            "anyOf": [
              {
                "type": "string"
              },
              {
                "type": "null"
              }
            ],
            "title": "Facility Name"
          },
          "facility_state": {
            "anyOf": [
              {
                "type": "string"
              },
              {
                "type": "null"
              }
            ],
            "title": "Facility State"
          },
          "inspection_date": {
            "anyOf": [
              {
                "type": "string",
                "format": "date"
              },
              {
                "type": "null"
              }
            ],
            "title": "Inspection Date"
          },
          "inspection_type": {
            "anyOf": [
              {
                "type": "string"
              },
              {
                "type": "null"
              }
            ],
            "title": "Inspection Type"
          },
          "inspector_name": {
            "anyOf": [
              {
                "type": "string"
              },
              {
                "type": "null"
              }
            ],
            "title": "Inspector Name"
          },
          "inspector_id": {
            "anyOf": [
              {
                "type": "string"
              },
              {
                "type": "null"
              }
            ],
            "title": "Inspector Id"
          },
          "violations_found": {
            "anyOf": [
              {
                "type": "boolean"
              },
              {
                "type": "null"
              }
            ],
            "title": "Violations Found"
          },
          "violation_count": {
            "anyOf": [
              {
                "type": "integer"
              },
              {
                "type": "null"
              }
            ],
            "title": "Violation Count"
          },
          "source_pdf": {
            "anyOf": [
              {
                "type": "string"
              },
              {
                "type": "null"
              }
            ],
            "title": "Source Pdf"
          },
          "source_pdf_path": {
            "anyOf": [
              {
                "type": "string"
              },
              {
                "type": "null"
              }
            ],
            "title": "Source Pdf Path"
          },
          "violations": {
            "items": {
              "$ref": "#/components/schemas/ViolationOut"
            },
            "type": "array",
            "title": "Violations"
          },
          "inventory": {
            "items": {
              "$ref": "#/components/schemas/InventoryOut"
            },
            "type": "array",
            "title": "Inventory"
          }
        },
        "type": "object",
        "required": [
          "id",
          "inspection_date",
          "inspection_type",
          "inspector_name",
          "inspector_id",
          "violations_found",
          "violation_count",
          "source_pdf",
          "source_pdf_path"
        ],
        "title": "InspectionOut"
      },
      "InspectorDetailOut": {
        "properties": {
          "inspector_id": {
            "anyOf": [
              {
                "type": "string"
              },
              {
                "type": "null"
              }
            ],
            "title": "Inspector Id"
          },
          "inspector_name": {
            "anyOf": [
              {
                "type": "string"
              },
              {
                "type": "null"
              }
            ],
            "title": "Inspector Name"
          },
          "total_inspections": {
            "type": "integer",
            "title": "Total Inspections"
          },
          "non_compliance_rate": {
            "type": "number",
            "title": "Non Compliance Rate"
          },
          "primary_state": {
            "anyOf": [
              {
                "type": "string"
              },
              {
                "type": "null"
              }
            ],
            "title": "Primary State"
          },
          "regional_average_rate": {
            "anyOf": [
              {
                "type": "number"
              },
              {
                "type": "null"
              }
            ],
            "title": "Regional Average Rate"
          },
          "anomaly_flag": {
            "type": "boolean",
            "title": "Anomaly Flag"
          },
          "inspections": {
            "items": {
              "$ref": "#/components/schemas/InspectionOut"
            },
            "type": "array",
            "title": "Inspections"
          }
        },
        "type": "object",
        "required": [
          "inspector_id",
          "inspector_name",
          "total_inspections",
          "non_compliance_rate",
          "primary_state",
          "regional_average_rate",
          "anomaly_flag"
        ],
        "title": "InspectorDetailOut"
      },
      "InspectorListOut": {
        "properties": {
          "total": {
            "type": "integer",
            "title": "Total"
          },
          "limit": {
            "type": "integer",
            "title": "Limit"
          },
          "offset": {
            "type": "integer",
            "title": "Offset"
          },
          "results": {
            "items": {
              "$ref": "#/components/schemas/InspectorOut"
            },
            "type": "array",
            "title": "Results"
          }
        },
        "type": "object",
        "required": [
          "total",
          "limit",
          "offset",
          "results"
        ],
        "title": "InspectorListOut"
      },
      "InspectorOut": {
        "properties": {
          "inspector_id": {
            "anyOf": [
              {
                "type": "string"
              },
              {
                "type": "null"
              }
            ],
            "title": "Inspector Id"
          },
          "inspector_name": {
            "anyOf": [
              {
                "type": "string"
              },
              {
                "type": "null"
              }
            ],
            "title": "Inspector Name"
          },
          "total_inspections": {
            "type": "integer",
            "title": "Total Inspections"
          },
          "non_compliance_rate": {
            "type": "number",
            "title": "Non Compliance Rate"
          },
          "primary_state": {
            "anyOf": [
              {
                "type": "string"
              },
              {
                "type": "null"
              }
            ],
            "title": "Primary State"
          },
          "regional_average_rate": {
            "anyOf": [
              {
                "type": "number"
              },
              {
                "type": "null"
              }
            ],
            "title": "Regional Average Rate"
          },
          "anomaly_flag": {
            "type": "boolean",
            "title": "Anomaly Flag"
          }
        },
        "type": "object",
        "required": [
          "inspector_id",
          "inspector_name",
          "total_inspections",
          "non_compliance_rate",
          "primary_state",
          "regional_average_rate",
          "anomaly_flag"
        ],
        "title": "InspectorOut"
      },
      "InventoryOut": {
        "properties": {
          "id": {
            "type": "integer",
            "title": "Id"
          },
          "scientific_name": {
            "anyOf": [
              {
                "type": "string"
              },
              {
                "type": "null"
              }
            ],
            "title": "Scientific Name"
          },
          "common_name": {
            "anyOf": [
              {
                "type": "string"
              },
              {
                "type": "null"
              }
            ],
            "title": "Common Name"
          },
          "count": {
            "anyOf": [
              {
                "type": "integer"
              },
              {
                "type": "null"
              }
            ],
            "title": "Count"
          },
          "source_pdf": {
            "anyOf": [
              {
                "type": "string"
              },
              {
                "type": "null"
              }
            ],
            "title": "Source Pdf"
          }
        },
        "type": "object",
        "required": [
          "id",
          "scientific_name",
          "common_name",
          "count",
          "source_pdf"
        ],
        "title": "InventoryOut"
      },
...
      "ValidationError": {
        "properties": {
          "loc": {
            "items": {
              "anyOf": [
                {
                  "type": "string"
                },
                {
                  "type": "integer"
                }
              ]
            },
            "type": "array",
            "title": "Location"
          },
          "msg": {
            "type": "string",
            "title": "Message"
          },
          "type": {
            "type": "string",
            "title": "Error Type"
          },
          "input": {
            "title": "Input"
          },
          "ctx": {
            "type": "object",
            "title": "Context"
          }
        },
        "type": "object",
        "required": [
          "loc",
          "msg",
          "type"
        ],
        "title": "ValidationError"
      },
      "ViolationOut": {
        "properties": {
          "id": {
            "type": "integer",
            "title": "Id"
          },
          "severity": {
            "anyOf": [
              {
                "type": "string"
              },
              {
                "type": "null"
              }
            ],
            "title": "Severity"
          },
          "section": {
            "anyOf": [
              {
                "type": "string"
              },
              {
                "type": "null"
              }
            ],
            "title": "Section"
          },
          "description": {
            "anyOf": [
              {
                "type": "string"
              },
              {
                "type": "null"
              }
            ],
            "title": "Description"
          },
          "source_pdf": {
            "anyOf": [
              {
                "type": "string"
              },
              {
                "type": "null"
              }
            ],
            "title": "Source Pdf"
          },
          "source_page": {
            "anyOf": [
              {
                "type": "integer"
              },
              {
                "type": "null"
              }
            ],
            "title": "Source Page"
          }
        },
        "type": "object",
        "required": [
          "id",
          "severity",
          "section",
          "description",
          "source_pdf",
          "source_page"
        ],
        "title": "ViolationOut"
      }
    }
  }
}